Allow zero or one args for :VimspectorEval (like :VimspectorWatch)
I looked for a test to cargo cult, but I couldn't find any place where
`:VimspectorWatch` was called with no args and fed a string to the
`input()` call.

I also looked at the commit that added the same behavior for
VimspectorWatch (4586aa36) and it included no tests. If there's an easy
way to test this, please let me know.