Add the ${lineNumber} predefined variable
When writing a debug profile configuration, it's useful to be able to
access the exact line in the source from which debugging is launched.
For example, this can be used to find the test function to debug.

VSCode already supports a ${lineNumber} predefined variable. Add support
for this variable in Vimspector.

Fixes puremourning/vimspector#404