Support readMemory request
You can hit <leader>m in the variables and watches window to dump a
section of memory associated with the variable under the cursor.

This uses an xxd-like output which should be familiar to most users.
There's no interactivity, you have to redo the request for different
offsets (for example) and you have to specify the length yourself.

This is really the MVP for this and there are a number of other possible
enhancements