Add basic session save/restore functions
Manual load/store of:
 - breakpoins
 - saved "entered data" in the debug session
 - watches

We wanted to automate this store/load but it's just really hard to know
when to do it. So we make it a manual operation for save/load specifying
a file name. We can still add something automatic if we can work out the
best way to do that.