Support arm64 runtime
We add <OS>_<Platform> blocks to the installer; these override the <OS>
and "all" blocks (in that order). This allows us to download different
binaries for different platforms, but for mostly the configuration to
remain per-OS.

This works on arm64 Mac and in the docker container on linux arm64 (sort
of; there's some bug where 'love' won't install in the container).