Delve: Make output visible
The delve-dap debug adapter doesn't send its output to the correct
channel via dap, rather prints to its own stdout. This behaviour is
buggy, but it seems that it's the job of every DAP client to fix it
rather than fixing it at source.

My solutin is to allow adapter specs to say that they need to be run in
a terminal rather than a background job. This would in theory allow for
stdin to work too.

An alternative, or perhaps complementary, solution would be to attach
the stdin/out of the 'channel-based server job' to a buffer and inlcude
that in the Output window.