Fix different break behaviour with gcc8
Gcc 8 (or perhaps gdb, unsure) changed something in the dbeug info such
that there is now a breakpoint target on the opening brace of a
function, thus the function breakpoitns (and cpptools stop on entry)
behave differnently from clang and indeed from gcc on arm.

So we have to hack the tests to know that they are different. Sigh.

We also make the manual test container disable retries and use a
particular base dir by default. this is for convenience.