install_gadgets: handle '~' correctly for base dir
The os.path.realpath function doesn't handle '~' symbols properly, so
when given the option '--basedir ~/something' in the cli, a mish mash of
directories are being generated; Preprocessing it with 'expanduser' does
the trick.

Added a 'strip' as well for good measure (extra spaces can
find their way in their while using this through, well, a debugger).

Signed-off-by: Omer Caspi <omer.caspi@gmail.com>