Add support for instruction breakpoints
Due to limitations of DAP, they are always "temporary" in practice
today. This is due to the lack of a consistent and reliable way to
determine their location after stepping.

Due to the endless faffery required otherwise, we hack line breakpoints
to contain both line and instruction breakpoints. Where line breakpoints
in a disassembly buffer are always treated as instruction breakpoints.
That's terrible, but it's the least change way to do it, and refactoring
this further is beyond human capability.

So for now, you can set some breakpoints in disassembly and they get
triggered. zoot.

---

Support 'presistenly' displaying instruction breakpoints (SORT OF)

We just compare the address of the breakpoint at the time it was created
to the address of a given instruction, then _hack_ the filename.

This won't work for disabling, and deleting, but this is step one. The
next step is to actually _nmove_ the breakpoint to the new buffer.

OR, probably preferably, stop using different buffers for each
disassembly output altogether.

---

Rationalise address-to-line calculation

Rather than using a buffer-per-memory-reference, we use a single buffer
for all disassembly. THis makes it easier to work with signs and things.

Now, we use the disassembly view to get the new line number whenever we
do a SignToLine. We accept that sometimes this wont work, so instruction
line breakpoints can have a line of 0 meaning that they are valid (and
may be VERIFIED) but do not relate to any currently displayed
buffer/line. THis is already true for breakpoints with no assocated
existing buffer, so it's not a massive stretch.

Finally, we can't persist instruction breakpoints across sessions, so we
clear them on server closedown. As noted, it _might_ be possible, but
would only work sometimes, so not that useful, probably.