Add support for instruction breakpoints
Due to limitations of DAP, they are always "temporary" in practice
today. This is due to the lack of a consistent and reliable way to
determine their location after stepping.

Due to the endless faffery required otherwise, we hack line breakpoints
to contain both line and instruction breakpoints. Where line breakpoints
in a disassembly buffer are always treated as instruction breakpoints.
That's terrible, but it's the least change way to do it, and refactoring
this further is beyond human capability.

So for now, you can set some breakpoints in disassembly and they get
triggered. zoot.