Support 'presistenly' displaying instruction breakpoints (SORT OF)
We just compare the address of the breakpoint at the time it was created
to the address of a given instruction, then _hack_ the filename.

This won't work for disabling, and deleting, but this is step one. The
next step is to actually _nmove_ the breakpoint to the new buffer.

OR, probably preferably, stop using different buffers for each
disassembly output altogether.