Rationalise address-to-line calculation
Rather than using a buffer-per-memory-reference, we use a single buffer
for all disassembly. THis makes it easier to work with signs and things.

Now, we use the disassembly view to get the new line number whenever we
do a SignToLine. We accept that sometimes this wont work, so instruction
line breakpoints can have a line of 0 meaning that they are valid (and
may be VERIFIED) but do not relate to any currently displayed
buffer/line. THis is already true for breakpoints with no assocated
existing buffer, so it's not a massive stretch.

Finally, we can't persist instruction breakpoints across sessions, so we
clear them on server closedown. As noted, it _might_ be possible, but
would only work sometimes, so not that useful, probably.