Only clear temporary breakpoints when hit
Previously setting a temp breakpoint on a line with an existing line bp
turned the line bp into a temp one, and that got deleted on hit.

Now, we always add a new temp bp and only delete that one.

Also, when setting e line bp that already exists, make sure to update
the server and call the callback.

Fixes #719