Add a simple built-in process picker
It's written in go because that's an easy way to make a portable binary
(we may ship these from CI in future). Ths dependency is a port of
`psutil` from python. THis package is not easily ported/vendored as it
has a compiled component, hence the use of go from python to use a port
of a python library.

Allow multiple header lines in SelectFromList (finally).