Expand references in calculus function arguments
This allows for the use of variables as arguments to functions.

In particular, the following is now possible:

     "configuration": {
       "request": "attach",
       "processId": "${PickProcess(\"${programName}\")}",
       "program": "$BUILD/platform/bin64/${programName}",
       "expresssions": "native"
     }

The result is that you're asked for the program _name_ and it
filters for that, and uses it as the binary.

Doesn't work for function-calls-inside-function-calls because of not
enough escaping levels for parameters, but that's probably OK. The above
is a real use case however, and it works.

This is mad, gnarly stuff which probably shouldn't be used, but
that's also true of the whole of this entire macro-templating
system, which honestly needs to be compeltely ditched!