Fix expanding variables after restart
We use the connection property of the result when expanding, so we need
to update it when we get a new result. Most of the time this will be NOP
but in the case of a variable which is carried over from a previous
session, it will change the first time we refresh it.