Connection to gadgets on IPv6-enabled host is failing
On hosts where IPv6 is enabled and preferred by the system, gadgets
might be listening to ::1 instead of 127.0.0.1. In those situations,
vimspector times out, unable to connect to the gadget even though
they've both agreed about which port to use.

Changing the default host value to 'localhost' makes it possible for
IPv6-enabled hosts to use the same default in vimspector than in the
gadgets.