Add support for puppet DSL
This is using Puppet's own DAP server in the puppet-editor-services
project.

Debugging puppet is a tiny bit weird.. it mostly means debugging the
catalog building that the puppet server would be operating.

It gets weirder in that to debug in a real situation you need to supply
variables and facts that would be in place for the host in question.

But it can nonetheless be done, and the DAP server permits stepping
through the catalog building steps through the code while examining
variables and resources as they are evaluated.