explicitly bind puppet-language-server to 127.0.0.1
vimspector connects explicitly to the ipv4 localhost address. This is
meant to avoid irregularities caused by using "localhost" which may end
up being ::1 for some hosts instead. Both vimspector and all the LSP
servers need to agree on the address family being used.

IPv6 has gained some more traction in the recent years but it's still
not reached enough people so that we could instead default to ::1 for
all.