Normalise path letters on Windows
paths on windows are not case-sensitive, and we're not really
good about equivalent paths with differing case. However, it
would be nice to believe that normpath() deals with that. Which,
of course, it doesn't.

For the most part though it works, with the frustrating exception
of drive letters. So we hack it to normalise drive letters to
uppercase, which prevents duplicate paths appearing in the breakpoints
lists.

Fixes #898